refactor(adapters): split osv_client.rs into mod/dto/cvss submodules
osv_client.rs (643 lines) mixed three distinct concerns: the HTTP client
and VulnerabilityRepository implementation, OSV wire-format DTOs, and a
self-contained CVSS v3.1 base-score calculator. Split into a submodule
directory:

- osv_client/mod.rs: OsvClient struct, HTTP methods, VulnerabilityRepository impl
- osv_client/dto.rs: all Osv*/DatabaseSpecific wire-format structs
- osv_client/cvss.rs: parse_cvss_score, parse_severity_string, and their tests

Pure file reorganization; no behavior change. Struct/field visibility
raised from private to pub(super) only where cross-file access within
the submodule is required.

Closes #713
